requirements.txt
Django~=4.2.9
python-dateutil~=2.8.2
icecream~=2.1.3
django-admin
django-bootstrap5
django-bootstrap-icons
django-tables2
django-simple-history~=3.4.0
tqdm~=4.66.1
django-environ